refactor qwen35 runner orchestration
