+++ crates/runner/src/tensor_bytes.rs
@@ -1,0 +1,20 @@
+pub fn bf16_bytes_to_f32(bytes: &[u8]) -> Vec<f32> {
+    bytes
+        .chunks_exact(2)
+        .map(|chunk| half::bf16::from_le_bytes([chunk[0], chunk[1]]).to_f32())
+        .collect()
+}
+
+pub fn f32_bytes_to_f32(bytes: &[u8]) -> Vec<f32> {
+    bytes
+        .chunks_exact(4)
+        .map(|chunk| f32::from_le_bytes([chunk[0], chunk[1], chunk[2], chunk[3]]))
+        .collect()
+}
+
+pub fn f32_to_bf16_bytes(values: impl IntoIterator<Item = f32>) -> Vec<u8> {
+    values
+        .into_iter()
+        .flat_map(|value| half::bf16::from_f32(value).to_le_bytes())
+        .collect()
+}
